refactor(protocol): abstract identity store into own module
